chore: Merge changes from (#98) with changes from (#86)
An automatic merge of PR "implement additional gauntlet command (#98)"
with PR "multisig instantiation (#86)" resulted in lots of stuff
either not compiling or not working.

This fixes the bugs introduced by the automatic merge, by manually merging
the two branches.  Mostly, needed to add categories for all of the new commands.