feature: upload and instantiation of multisig contracts (#83)
* feature: upload & deploy cw4_group and cw3_flex_multisig (#83)

Instructions for uploading & deploying these 2 contracts have been added
to gauntlet-terra-contracts README.md, along with functionality.

This does not include ability to make multisig proposals and vote on
them--that will be in a separate PR.

This also includes support for downloading wasm artifacts from cwplus repo

* refactor command vs contract inputs (#83)

* wallet input refactor (#83)

* bugfix: Fix base64 encoding in downloader, remove unused params (#83)

It looks like downloading remote wasm files and uploading them was
never actually working (even on main).  The terra network was rejecting
the contract uploads due to them not being base64 encoded.  I think
the toString('base64') operation on a string must have been a no-op.
Seems to work after this change.

Other changes:
- Separate default versions for cwplus contracts and chainlink-terra contracts
- Remove unused params
- Workaround for exceptional naming of the execute shcema for cw20_base
  (cw20_execute.json instead of execute.json)

* chore: rebase to latest main, and update gauntlet README.md (#83)

* chore: validate that owners.length != 0, and improve error reporting (#83)

* chore: Merge changes from (#98) with changes from (#86)

An automatic merge of PR "implement additional gauntlet command (#98)"
with PR "multisig instantiation (#86)" resulted in lots of stuff
either not compiling or not working.

This fixes the bugs introduced by the automatic merge, by manually merging
the two branches.  Mostly, needed to add categories for all of the new commands.

Co-authored-by: RodrigoAD <15104916+RodrigoAD@users.noreply.github.com>