Load instances after reading network env file
This fixes a bug due to the instances were being loaded too soon.
Contract addresses were properly formatted only if they appeared
in .env; now they should also work if only defined in network/.env-*

Also, replaces this.addresses and this.address in Contract with
this.instances, including instance-specific names for each contract
to print instead of the type of contract.